pyproject.toml
[project]
name = "hayakoe"
version = "0.0.6"
description = "HayaKoe — Fast ONNX-optimized Japanese TTS based on Style-Bert-VITS2"
readme = "README.md"
requires-python = ">=3.10"
license = "AGPL-3.0"
authors = [
    { name = "LemonDouble" },
]

dependencies = [
    # Core
    "transformers>=5.0",
    "numpy>=2.0",
    "pydantic==2.12.5",
    "loguru==0.7.3",
    "huggingface-hub>=0.20",
    "onnxruntime>=1.17",
    # NLP - Japanese
    "lemon-pyopenjtalk-prebuilt==0.0.7",
    "pyworld-prebuilt==0.3.5.post2",
    "num2words==0.5.14",
    "numba==0.65.0",
]

[project.optional-dependencies]
gpu = [
    # torch는 CUDA 버전에 따라 별도 설치 필요:
    #   pip install torch --index-url https://download.pytorch.org/whl/cu126
    "safetensors>=0.4",
]

[project.urls]
Source = "https://github.com/LemonDouble/hayakoe"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["hayakoe"]

[tool.hatch.build.targets.sdist]
only-include = [
    "hayakoe",
    "LGPL_LICENSE",
    "LICENSE",
    "pyproject.toml",
    "README.md",
]

[tool.ruff]
extend-select = ["I"]

[tool.ruff.lint.isort]
lines-after-imports = 2

[dependency-groups]
dev = [
    # 전처리 서버
    "fastapi>=0.135.3",
    "python-multipart>=0.0.24",
    "uvicorn>=0.44.0",
    "soundfile>=0.13.1",
    # 전처리 ML (torch는 CUDA 버전에 맞게 별도 설치)
    "librosa>=0.11",
    "openai-whisper>=20240930",
    "audio-separator>=0.29",
    # 학습
    "tensorboard>=2.20.0",
    "torchaudio>=2.11.0",
    "pyannote-audio>=3.3",
    # CLI
    "typer>=0.15",
    "InquirerPy>=0.3.4",
    "rich>=13.0",
]
